locate playerstart position, portal stone in dungeon

+++ Assets/Scripts/Items/ItemInfoSO.cs
@@ -12,7 +12,7 @@ public class ItemInfoSO : ScriptableObject
 public class ConsumableItemInfoSO : ItemInfoSO
 {
     public EConsumableType consumableType;
-    public float amount;
+    public int amount;
 }
 
 [CreateAssetMenu(fileName = "EquipmentInfoSO", menuName = "ScriptableObjects/EquipmentInfoSO")]
@@ -25,14 +25,14 @@ public class EquipmentInfoSO : ItemInfoSO
 public class HeadInfoSO : EquipmentInfoSO
 {
     public EWeaponType headType;
-    public float armor;
+    public int armor;
 }
 
 [CreateAssetMenu(fileName = "WeaponInfoSO", menuName = "ScriptableObjects/WeaponInfoSO")]
 public class WeaponInfoSO : EquipmentInfoSO
 {
     public EWeaponType weaponType;
-    public float damage;
+    public int damage;
 }
 
 [CreateAssetMenu(fileName = "ArmorInfoSO", menuName = "ScriptableObjects/ArmorInfoSO")]
@@ -41,12 +41,12 @@ public class ArmorInfoSO : EquipmentInfoSO
     public EArmorType armorType;
     public Sprite itemSprite2;
     public Sprite itemSprite3;
-    public float armor;
+    public int armor;
 }
 
 [CreateAssetMenu(fileName = "ShieldInfoSO", menuName = "ScriptableObjects/ShieldInfoSO")]
 public class ShieldInfoSO : EquipmentInfoSO
 {
     public EShieldType shieldType;
-    public float armor;
+    public int armor;
 }
